+++ laktory/models/datasources/eventdatasource.py
@@ -28,7 +28,7 @@ class EventDataSource(BaseDataSource, DataEventHeader):
     fmt: Literal[FORMATS] = "JSON"
     multiline: bool = False
     header: bool = True
-    delimiter: str = None
+    read_options: dict[str, str] = {}
 
     # ----------------------------------------------------------------------- #
     # Readers                                                                 #
@@ -65,8 +65,8 @@ class EventDataSource(BaseDataSource, DataEventHeader):
             .option("recursiveFileLookup", True)
             .option("header", self.header)  # only apply to CSV format
         )
-        if self.delimiter is not None:
-            reader = reader.option("delimiter", self.delimiter)  # only apply to CSV format
+        if self.read_options:
+            reader = reader.options(self.read_options)
 
         # Load
         df = reader.load(self.event_root)