+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## [0.0.20] - 2023-11-24
 ###Added
-* `delimiter` option when reading CSV event data source
 * `header` option when reading CSV event data source 
+* `read_options` option when reading event data source
 
 ## [0.0.19] - 2023-11-23
 ###Fixed